requirements.txt
streamlit
pandas
plotly
openai
python-dotenv
openpyxl
python-docx
matplotlib
